runtime: Make runtime a separately compiled static library
This prevents us from compiling all of Format, String, File, etc along
with every single Jakt executable.
