+++ cmake/jakt-executable.cmake
@@ -32,6 +32,15 @@ function(add_jakt_compiler_flags target)
     target_compile_options("${target}" PRIVATE -Wa,-mbig-obj)
   endif()
   target_compile_features("${target}" PRIVATE cxx_std_20)
+
+  # Default to /MT and /MTd instead of /MD and /MDd to match what happens when invoking clang.exe manually
+  if (WIN32)
+    cmake_policy(GET CMP0091 msvc_runtime_prop_enabled)
+    if (NOT msvc_runtime_prop_enabled STREQUAL "NEW")
+      message(WARNING "CMake Policy CMP0091 is not set to NEW, linker errors from jakt_runtime may result")
+    endif()
+    set_property(TARGET "${target}" PROPERTY MSVC_RUNTIME_LIBRARY "MultiThreaded$<$<CONFIG:Debug>:Debug>")
+  endif()
 endfunction()
 
 function(add_jakt_executable executable)
@@ -41,7 +50,7 @@ function(add_jakt_executable executable)
   set(cpp_output "${executable}_${main_base}.cpp")
 
   if (NOT JAKT_EXECUTABLE_COMPILER)
-    set(JAKT_EXECUTABLE_COMPILER Jakt::jakt_stage1)
+    set(JAKT_EXECUTABLE_COMPILER Jakt::jakt)
   endif()
 
   if (NOT JAKT_EXECUTABLE_RUNTIME_DIRECTORY)
@@ -66,6 +75,7 @@ function(add_jakt_executable executable)
 
   add_executable("${executable}" "${cpp_output}")
   target_include_directories("${executable}" PRIVATE "$<BUILD_INTERFACE:${JAKT_EXECUTABLE_RUNTIME_DIRECTORY}>")
+  target_link_libraries("${executable}" PRIVATE Jakt::jakt_runtime)
   add_dependencies("${executable}" "generate_${executable}")
   add_jakt_compiler_flags("${executable}")
 endfunction()
